Fix circular imports
- Fixed circular import problems by moving like-classes into one
  file as a quick fix
