--- src/chat/models/unauthorizedChatRoomMember.js
@@ -1,13 +1,0 @@
-const ChatRoomMember = require('./chatRoomMember');
-
-class UnauthorizedChatRoomMember extends ChatRoomMember {
-  constructor(username, role, listingId) {
-    super();
-
-    this.username = username;
-    this.role = role;
-    this.listingId = listingId;
-  }
-};
-
-module.exports = UnauthorizedChatRoomMember;
